refactor(web): move trace detail URL builder
